fix(cluster/talos): Add k8s endpoint health check
The windsor command now supports including k8s api endpoints in its node health check. Leveraging this may help prevent issues when the cluster nodes are healthy but the API isn't yet accessible.

In order to use this routine, the kubeconfig must have been created. It makes sense to perform the kubeconfig generation inside the machine module when `bootstrap = true`.

Also sets parallelism to 1 in the template when building the talos cluster. This will be leveraged in the future for upgrades, but is appropriate to establish as part of the sequenced health checking routine.